// src/app/heroes/interfaces/hero.interface.ts
// Generated by https://quicktype.io

export interface Hero {
    id:             string;
    name:           string;
    localized_name: string;
    primary_attr:   PrimaryAttr;
    attack_type:    AttackType;
    roles:          Role[];
    legs:           number | null;
    alt_image?:     string;
}

export enum AttackType {
    Melee = "Melee",
    Ranged = "Ranged",
}

export enum PrimaryAttr {
    Agi = "agi",
    All = "all",
    Int = "int",
    Str = "str",
}

export enum Role {
    Carry = "Carry",
    Disabler = "Disabler",
    Durable = "Durable",
    Escape = "Escape",
    Initiator = "Initiator",
    Nuker = "Nuker",
    Pusher = "Pusher",
    Support = "Support",
}